Pull a bunch of specs over to to AbstractCustomField
